requirements.txt
asgiref==3.7.2
asttokens==2.2.1
backcall==0.2.0
decorator==5.1.1
Django==4.2.3
django-autocomplete-light==3.9.7
django-filter==23.2
django-psdb-engine==1.0.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
et-xmlfile==1.1.0
executing==1.2.0
gunicorn==21.2.0
ipython==8.14.0
jedi==0.19.0
load-dotenv==0.1.0
Markdown==3.4.4
matplotlib-inline==0.1.6
mysqlclient==2.2.0
numpy==1.25.1
openpyxl==3.1.2
packaging==23.1
pandas==2.0.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.39
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.16.1
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
traitlets==5.9.0
tzdata==2023.3
validate_email==1.3
wcwidth==0.2.6
whitenoise==6.5.0
